Feat: 계정 삭제 API URI 수정
