+++ core/src/main/java/org/bookwoori/core/domain/member/controller/MemberController.java
@@ -6,6 +6,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bookwoori.core.domain.member.dto.request.UpdateMemberRequestDto;
+import org.bookwoori.core.domain.member.facade.AuthFacade;
 import org.bookwoori.core.domain.member.facade.MemberFacade;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -24,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberController {
 
     private final MemberFacade memberFacade;
+    private final AuthFacade authFacade;
 
     @Operation(summary = "프로필 수정", description = "닉네임, 프로필 이미지를 수정합니다.")
     @PatchMapping(value = "/me", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -43,5 +45,12 @@ public class MemberController {
     @GetMapping(value = "/me")
     public ResponseEntity<?> getMyProfile() {
         return ResponseEntity.ok(memberFacade.getMyProfile());
+    }
+
+    @Operation(summary = "계정 삭제", description = "회원 상태를 INACTIVE로 변경하고 닉네임을 '(알 수 없음)'으로 변경합니다.")
+    @PatchMapping("/delete")
+    public ResponseEntity<?> deleteMember() {
+        authFacade.deleteMember();
+        return ResponseEntity.ok().build();
     }
 }
